+ Рефакторинг ProfileViewController под MVC
