+++ FakeNFT/Scenes /Profile/ProfileView.swift
@@ -2,31 +2,15 @@
 //  ProfileView.swift
 //  FakeNFT
 //
-//  Created by Anton Demidenko on 11.11.24..
+//  Created by Anton Demidenko on 12.11.24..
 //
-
-import Foundation
 import UIKit
-import WebKit
-import Kingfisher
-import ProgressHUD
 
-final class ProfileView: UIViewController {
+final class ProfileView: UIView {
     
-    private let servicesAssembly: ServicesAssembly
+    var websiteLabelTapped: ((String) -> Void)?
     
     // MARK: - UI Elements
-    private lazy var editProfileButton: UIButton = {
-        let button = UIButton(type: .system)
-        
-        let configuration = UIImage.SymbolConfiguration(pointSize: 20, weight: .bold)
-        let image = UIImage(systemName: "square.and.pencil", withConfiguration: configuration)
-        
-        button.setImage(image, for: .normal)
-        button.tintColor = UIColor(named: "YBlackColor")
-        button.addTarget(self, action: #selector(editProfileTapped), for: .touchUpInside)
-        return button
-    }()
     
     private lazy var profileAvatar: UIImageView = {
         let imageView = UIImageView()
@@ -58,11 +42,6 @@ final class ProfileView: UIViewController {
         return label
     }()
     
-    private lazy var webView: WKWebView = {
-        let webView = WKWebView()
-        return webView
-    }()
-    
     private lazy var profileInfoLabel: UILabel = {
         let label = UILabel()
         label.text = NSLocalizedString("NoInformation", comment: "")
@@ -75,6 +54,7 @@ final class ProfileView: UIViewController {
     
     private lazy var profileTableView: UITableView = {
         let tableView = UITableView()
+        tableView.delegate = self
         tableView.dataSource = self
         tableView.isScrollEnabled = false
         tableView.rowHeight = 54
@@ -89,47 +69,23 @@ final class ProfileView: UIViewController {
     }()
     
     // MARK: - Initialization
-    init(servicesAssembly: ServicesAssembly) {
-        self.servicesAssembly = servicesAssembly
-        super.init(nibName: nil, bundle: nil)
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setupLayout()
     }
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
     
-    // MARK: - Lifecycle
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        view.backgroundColor = .systemBackground
-        setupNavigationBar()
-        setupLayout()
-        ProgressHUD.show()
-        loadProfile()
-    }
-    
-    // MARK: - Setup Methods
-    private func setupNavigationBar() {
-        navigationItem.rightBarButtonItem = UIBarButtonItem(customView: editProfileButton)
-        editProfileButton.translatesAutoresizingMaskIntoConstraints = false
-        
-        NSLayoutConstraint.activate([
-            editProfileButton.widthAnchor.constraint(equalToConstant: 44),
-            editProfileButton.heightAnchor.constraint(equalToConstant: 44),
-            editProfileButton.trailingAnchor.constraint(equalTo: navigationItem.rightBarButtonItem!.customView!.trailingAnchor, constant: -9),
-            editProfileButton.centerYAnchor.constraint(equalTo: navigationItem.rightBarButtonItem!.customView!.centerYAnchor)
-        ])
-    }
-    
     private func setupLayout() {
         
-        view.addSubview(profileContainerView)
+        addSubview(profileContainerView)
         profileContainerView.addSubview(profileAvatar)
         profileContainerView.addSubview(userNameLabel)
         profileContainerView.addSubview(profileInfoLabel)
         profileContainerView.addSubview(userWebSiteLabel)
-        view.addSubview(profileTableView)
+        addSubview(profileTableView)
         
         profileContainerView.translatesAutoresizingMaskIntoConstraints = false
         profileAvatar.translatesAutoresizingMaskIntoConstraints = false
@@ -139,9 +95,9 @@ final class ProfileView: UIViewController {
         profileTableView.translatesAutoresizingMaskIntoConstraints = false
         
         NSLayoutConstraint.activate([
-            profileContainerView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 20),
-            profileContainerView.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 16),
-            profileContainerView.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -16)
+            profileContainerView.topAnchor.constraint(equalTo: safeAreaLayoutGuide.topAnchor, constant: 20),
+            profileContainerView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 16),
+            profileContainerView.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -16)
         ])
         
         NSLayoutConstraint.activate([
@@ -166,13 +122,19 @@ final class ProfileView: UIViewController {
         
         NSLayoutConstraint.activate([
             profileTableView.topAnchor.constraint(equalTo: profileContainerView.bottomAnchor, constant: 40),
-            profileTableView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            profileTableView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            profileTableView.leadingAnchor.constraint(equalTo: leadingAnchor),
+            profileTableView.trailingAnchor.constraint(equalTo: trailingAnchor),
             profileTableView.heightAnchor.constraint(equalToConstant: 54 * 3)
         ])
     }
     
-    private func updateUI(with profile: Profile) {
+    @objc private func didTapOnWebsiteLabel() {
+        if let text = userWebSiteLabel.text {
+            websiteLabelTapped?(text)
+        }
+    }
+    
+    func updateUI(with profile: Profile) {
         userNameLabel.text = profile.name
         
         if let avatarURLString = profile.avatar, let url = URL(string: avatarURLString) {
@@ -189,36 +151,11 @@ final class ProfileView: UIViewController {
         
         profileInfoLabel.text = profile.description ?? NSLocalizedString("NoInformation", comment: "")
     }
-    
-    // MARK: - Actions
-    
-    @objc private func editProfileTapped() {
-    }
-    
-    @objc private func didTapOnWebsiteLabel() {
-        guard let urlString = userWebSiteLabel.text,
-              let url = URL(string: urlString) else { return }
-        
-        let request = URLRequest(url: url)
-        webView.load(request)
-        
-        let webViewController = UIViewController()
-        webViewController.view.addSubview(webView)
-        webView.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            webView.topAnchor.constraint(equalTo: webViewController.view.topAnchor),
-            webView.bottomAnchor.constraint(equalTo: webViewController.view.bottomAnchor),
-            webView.leadingAnchor.constraint(equalTo: webViewController.view.leadingAnchor),
-            webView.trailingAnchor.constraint(equalTo: webViewController.view.trailingAnchor)
-        ])
-        
-        webViewController.modalPresentationStyle = .pageSheet
-        present(webViewController, animated: true, completion: nil)
-    }
 }
 
 // MARK: - UITableViewDataSource
-extension ProfileView: UITableViewDataSource {
+
+extension ProfileView: UITableViewDelegate, UITableViewDataSource {
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return 3
     }
@@ -245,36 +182,15 @@ extension ProfileView: UITableViewDataSource {
         
         cell.accessoryView = chevronImageView
         cell.tintColor = UIColor(named: "YBlackColor")
+        cell.selectionStyle = .none
         
         
         return cell
     }
-}
-
-extension ProfileView {
-    private func loadProfile() {
-        servicesAssembly.profileService.loadProfile { [weak self] result in
-            DispatchQueue.main.async {
-                ProgressHUD.dismiss()
-                switch result {
-                case .success(let profile):
-                    self?.updateUI(with: profile)
-                    print("\(profile.name), \(profile.avatar), \(profile.website)")
-                case .failure(let error):
-                    print("error \(error)")
-                }
-            }
-        }
-    }
-}
-
-extension UIImageView {
-    func loadImage(from url: URL) {
-        URLSession.shared.dataTask(with: url) { data, _, _ in
-            guard let data = data else { return }
-            DispatchQueue.main.async {
-                self.image = UIImage(data: data)
-            }
-        }.resume()
+    
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        tableView.deselectRow(at: indexPath, animated: true)
+        
+        print("Ячейка выбрана")
     }
 }
